+++ src/clone-node.ts
@@ -13,7 +13,10 @@ async function cloneCanvasElement(canvas: HTMLCanvasElement) {
 }
 
 async function cloneVideoElement(video: HTMLVideoElement, options: Options) {
-  if (video.currentSrc) {
+  if (
+    (video.currentSrc || video.srcObject) &&
+    (video.poster == null || video.poster === '' || video.currentTime > 0)
+  ) {
     const canvas = document.createElement('canvas')
     const ctx = canvas.getContext('2d')
     canvas.width = video.clientWidth
@@ -61,6 +64,26 @@ async function cloneSingleNode<T extends HTMLElement>(
     return cloneIFrameElement(node)
   }
 
+  // Manage image with relative path in iframe
+  const src = (node as any).src
+  if (src != null) {
+    // Check if the image is in the top window
+    if (
+      node.ownerDocument.defaultView &&
+      window.top &&
+      node.ownerDocument.defaultView.location.host !== window.top.location.host
+    ) {
+      const baseUrl = `${node.ownerDocument.defaultView.location.protocol}//${node.ownerDocument.defaultView.location.host}`
+      const clonedNode = node.cloneNode(false) as any
+      if (!src.startsWith('https:')) {
+        clonedNode.src = `${baseUrl}${src}`
+      } else {
+        clonedNode.src = `${src}`
+      }
+      return clonedNode as T
+    }
+  }
+
   return node.cloneNode(false) as T
 }
 
@@ -74,16 +97,37 @@ async function cloneChildren<T extends HTMLElement>(
 ): Promise<T> {
   let children: T[] = []
 
-  if (isSlotElement(nativeNode) && nativeNode.assignedNodes) {
+  if (
+    isSlotElement(nativeNode) &&
+    nativeNode.assignedNodes &&
+    nativeNode.assignedNodes().length > 0
+  ) {
     children = toArray<T>(nativeNode.assignedNodes())
   } else if (
-    isInstanceOfElement(nativeNode, HTMLIFrameElement) &&
-    nativeNode.contentDocument?.body
+    nativeNode.parentElement &&
+    nativeNode.parentElement.hasAttribute('pdf-document') &&
+    !nativeNode.parentElement.classList.contains('book-container')
   ) {
-    children = toArray<T>(nativeNode.contentDocument.body.childNodes)
+    // Snapshot only one page of a multi-page PDF
+    children = toArray<T>(
+      (nativeNode.shadowRoot ?? nativeNode).childNodes,
+    ).filter((child) => child.getAttribute('isCurrentPage') === 'true')
+  } else if (
+    isInstanceOfElement(nativeNode, HTMLIFrameElement)
+    // && nativeNode.contentDocument?.body
+  ) {
+    // Iframe's children are already cloned in CloneSingleNode function
+    children = [] // toArray<T>(nativeNode.contentDocument.body.childNodes)
   } else {
     children = toArray<T>((nativeNode.shadowRoot ?? nativeNode).childNodes)
   }
+
+  // Keep only visible elements and delete source tags
+  children = children.filter(
+    (child) =>
+      (child.style == null || child.style.display !== 'none') &&
+      (child.tagName == null || child.tagName.toLowerCase() !== 'source'),
+  )
 
   if (
     children.length === 0 ||
@@ -174,6 +218,9 @@ function cloneScrollbarPositions<T extends HTMLElement>(
   nativeNode: T,
   clonedNode: T,
 ) {
+  // Snapshot only on page of a multi-page PDF so no need to scroll
+  if (nativeNode.hasAttribute('pdf-document')) return
+
   if (nativeNode.scrollLeft !== 0 || nativeNode.scrollTop !== 0) {
     for (let i = 0; i < clonedNode.children.length; i += 1) {
       const child = clonedNode.childNodes[i] as any as HTMLSelectElement